+++ addon/components/yeti-table/foot/row/cell/component.js
@@ -1,6 +1,7 @@
 import Component from '@ember/component';
 
 import { tagName } from '@ember-decorators/component';
+import { reads } from '@ember/object/computed';
 
 import layout from './template';
 
@@ -19,12 +20,20 @@ import layout from './template';
 
  */
 @tagName('')
-class Cell extends Component {
+class FootCell extends Component {
   layout = layout;
 
   theme;
 
   parent;
+
+  /**
+   * Controls the visibility of the current cell. Keep in mind that this property
+   * won't just hide the column using css. The DOM for the column will be removed.
+   * Defaults to the `visible` argument of the corresponding column.
+   */
+  @reads('column.visible')
+  visible;
 
   init() {
     super.init(...arguments);
@@ -42,4 +51,4 @@ class Cell extends Component {
   }
 }
 
-export default Cell;
+export default FootCell;