+++ tests/integration/components/yeti-table/theme-test.js
@@ -49,7 +49,7 @@ module('Integration | Component | yeti-table (theme)', function(hooks) {
 
   test('renders table with correct theme (header)', async function (assert) {
     await render(hbs`
-      <YetiTable @data={{data}} @theme={{theme}} as |table|>
+      <YetiTable @data={{this.data}} @theme={{theme}} as |table|>
 
         <table.header as |header|>
           <header.column @prop="firstName">
@@ -78,7 +78,7 @@ module('Integration | Component | yeti-table (theme)', function(hooks) {
 
   test('renders table with correct theme (head)', async function (assert) {
     await render(hbs`
-      <YetiTable @data={{data}} @theme={{theme}} as |table|>
+      <YetiTable @data={{this.data}} @theme={{theme}} as |table|>
 
         <table.head as |head|>
           <head.row as |row|>
@@ -138,7 +138,7 @@ module('Integration | Component | yeti-table (theme)', function(hooks) {
     };
 
     await render(hbs`
-      <YetiTable @data={{data}} @theme={{theme}} as |table|>
+      <YetiTable @data={{this.data}} @theme={{theme}} as |table|>
 
         <table.head as |head|>
           <head.row as |row|>